Add redux-thunk, refactor successful order handler
